--- src/main/java/com/example/LectureServerApplication.java
@@ -1,13 +1,0 @@
-package com.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LectureServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LectureServerApplication.class, args);
-	}
-
-}
