--- src/main/java/com/example/global/exception/ClientException.java
@@ -1,12 +1,0 @@
-package com.example.global.exception;
-
-import lombok.Getter;
-
-@Getter
-public class ClientException extends RuntimeException {
-
-    public ClientException(String message) {
-        super(message);
-    }
-
-}
