--- src/main/java/com/example/global/exception/DuplicateException.java
@@ -1,10 +1,0 @@
-package com.example.global.exception;
-
-import lombok.Getter;
-
-@Getter
-public class DuplicateException extends RuntimeException{
-    public DuplicateException(String message) {
-        super(message);
-    }
-}
