--- src/main/java/com/example/global/exception/JwtException.java
@@ -1,10 +1,0 @@
-package com.example.global.exception;
-
-import lombok.Getter;
-
-@Getter
-public class JwtException extends RuntimeException{
-    public JwtException(String message) {
-        super(message);
-    }
-}
