--- src/main/java/com/example/global/exception/MethodException.java
@@ -1,11 +1,0 @@
-package com.example.global.exception;
-
-import lombok.Getter;
-
-@Getter
-public class MethodException extends RuntimeException{
-    public MethodException(String message) {
-        super(message);
-    }
-
-}
