--- src/main/java/com/example/global/exception/NotFoundException.java
@@ -1,12 +1,0 @@
-package com.example.global.exception;
-
-import lombok.Getter;
-
-@Getter
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-
-}
-
