--- src/main/java/com/example/lecture/domain/entity/Semester.java
@@ -1,5 +1,0 @@
-package com.example.lecture.domain.entity;
-
-public enum Semester {
-    FIRST,SECOND
-}
