--- src/main/java/com/example/lecture/domain/entity/Status.java
@@ -1,5 +1,0 @@
-package com.example.lecture.domain.entity;
-
-public enum Status {
-    ACCEPT,DENIED,HOLDING
-}
