+++ src/main/java/com/lms/example/LectureServerApplication.java
@@ -1,0 +1,13 @@
+package com.lms.example;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LectureServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LectureServerApplication.class, args);
+	}
+
+}
