+++ src/main/java/com/lms/example/LectureTest.sql
@@ -1,0 +1,6 @@
+#강의 승인
+UPDATE lecture.lecture
+SET status = 'ACCEPT'
+WHERE id IN (6,7);
+
+
