+++ src/main/java/com/lms/example/global/exception/ClientException.java
@@ -1,0 +1,12 @@
+package com.lms.example.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class ClientException extends RuntimeException {
+
+    public ClientException(String message) {
+        super(message);
+    }
+
+}
