+++ src/main/java/com/lms/example/global/exception/DuplicateException.java
@@ -1,0 +1,10 @@
+package com.lms.example.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class DuplicateException extends RuntimeException{
+    public DuplicateException(String message) {
+        super(message);
+    }
+}
