+++ src/main/java/com/lms/example/global/exception/JwtException.java
@@ -1,0 +1,10 @@
+package com.lms.example.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class JwtException extends RuntimeException{
+    public JwtException(String message) {
+        super(message);
+    }
+}
