+++ src/main/java/com/lms/example/global/exception/MethodException.java
@@ -1,0 +1,11 @@
+package com.lms.example.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class MethodException extends RuntimeException{
+    public MethodException(String message) {
+        super(message);
+    }
+
+}
