+++ src/main/java/com/lms/example/global/exception/NotFoundException.java
@@ -1,0 +1,12 @@
+package com.lms.example.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+
+}
+
