+++ src/main/java/com/lms/example/global/kafka/KafkaAction.java
@@ -1,0 +1,5 @@
+package com.lms.example.global.kafka;
+
+public enum KafkaAction {
+    CREATE,UPDATE,DELETE
+}
