+++ src/main/java/com/lms/example/global/kafka/KafkaProducer.java
@@ -1,0 +1,30 @@
+package com.lms.example.global.kafka;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class KafkaProducer {
+
+    private final KafkaTemplate<String, String> kafkaTemplate;
+
+    public KafkaLecture saveLecture(String topic, KafkaLecture kafkaMember) {
+        ObjectMapper mapper = new ObjectMapper();
+        String jsonInString = "";
+
+        try {
+            jsonInString = mapper.writeValueAsString(kafkaMember);
+        } catch (JsonProcessingException ex) {
+            ex.printStackTrace();
+        }
+
+        kafkaTemplate.send(topic, jsonInString);
+        return kafkaMember;
+    }
+
+}
+
