+++ src/main/java/com/lms/example/lecture/domain/entity/Semester.java
@@ -1,0 +1,5 @@
+package com.lms.example.lecture.domain.entity;
+
+public enum Semester {
+    FIRST,SECOND
+}
