+++ src/main/java/com/lms/example/lecture/domain/entity/Status.java
@@ -1,0 +1,5 @@
+package com.lms.example.lecture.domain.entity;
+
+public enum Status {
+    ACCEPT,DENIED,HOLDING
+}
