+++ src/main/java/com/lms/example/lecture/dto/request/AdminDenyLectureRequest.java
@@ -1,0 +1,26 @@
+package com.lms.example.lecture.dto.request;
+
+import com.lms.example.lecture.domain.entity.Lecture;
+import com.lms.example.lecture.domain.entity.Status;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class AdminDenyLectureRequest {
+    private Long lectureId;
+
+    public Lecture toEntity() {
+        return Lecture.builder()
+                .id(lectureId)
+                .status(Status.DENIED)
+                .acceptedAt(LocalDateTime.now())
+                .build();
+    }
+}
