+++ src/main/resources/application.yml
@@ -1,0 +1,23 @@
+server:
+  port: 1002
+
+spring:
+  application:
+    name: LECTURE-SERVER
+
+  datasource:
+    url: jdbc:mysql://localhost:3306/lecture?serverTimezone=Asia/Seoul
+    username: root
+    password: 1q2w3e4r!!
+  jpa:
+    hibernate:
+      ddl-auto: create
+    show-sql: true
+
+eureka:
+  client:
+    service-url:
+      defaultZone: http://localhost:8761/eureka
+  instance:
+    prefer-ip-address: true
+    instanceId: ${spring.application.name}:${vcap.application.instance_id:${spring.application.instance_id:${random.value}}}
